Refactor Canvas Out of Example
Rebuild Canvas into src library via TDD, one test at a time.

I only moved over things that I felt would be useful and stable.

I did not move over any methods that write/draw/modify the Canvas,
with the exception of `resize` and `clear`, in an attempt to keep
the Canvas class clean and pure.

The reason for this is I still haven't discovered what the interface
between Shapes and the Canvas will be. I don't want the Canvas to know
about every possible Shape. Shapes must have a generic interface the
Canvas can understand. Otherwise the Canvas will become bloated and
inflexible. And Shapes musn't know about the Canvas otherwise
they wouldn't be portable to other targets.

This is going to take some time to figure out. So I'm going to hold
off moving anything into the src library that I'm not confident will
be stable and useful.