Pull Canvas Component Out of Example
This will be useful for building other examples. It will be useful
for figuring out the Canvas.draw(shapes) api as well. We'll pass a
list of shapes as a prop into the component from the state of the
parent, demonstrating how the shapes are persistent state and how
they are input to rendering.