Add ResizeSensor Component
The window resizing does not always mean the canvas will resize. And
the canvas can resize independent of the window. Therefore, we don't
give a shit about the window resizing.

We really want to redraw the canvas when the dimensions of the canvas
itself change.

So we create a little component called ResizeSensor and smack it into
our Canvas component to draw when element resize is detected.

The way it works is we lay an iframe under the canvas then listen to
it's contentWindow resize event to trigger redraw.

We had to wrap the canvas with a wrapper div since nesting the iframe
inside the canvas results in the iframe not having a width and height.

We move the styles up from the canvas element to the wrapper, so that
the dimensions of the cavnas component can be controlled on a single
element, the wrapper, and the canvas and iframe fill the wrapper.

Pretty slick if I do say so myself. Just slap a ResizeSensor in any
element and you can respond to any changes in it's dimensions. Great
for responsive components.