Refactoring
Make the height of the canvas component proportionate to the width using
the padding top trick (percent value for padding is evaluated as a percent
of the width). Aspect ratio is determined based on screen, which is
conducive to full screen, which seems appropriate since the primary application
of 3D programming is gaming.

Remove background color on ResizeSensor.