Perspective/Projection
To achieve perspective, I like to think of suspending the objects in
a frustum shaped jello, then compressing the jello frustum into a
cuboid shape, causing far side of each object to also be compressed.

It's a matter of determining how far between two sides of the frustum
a point is, then porting that relative point to a point on the canvas.

Point -> percent -> canvas.